refactor: split postgresql database service
